requirements.txt
aiodns==1.1.1
aiohttp==3.5.4
apturl==0.5.2
asn1crypto==0.24.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
Brlapi==0.6.6
ccxt==1.18.999
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
command-not-found==0.3
cryptography==2.7
cupshelpers==1.0
cycler==0.10.0
decorator==4.4.0
defer==1.0.6
defusedxml==0.6.0
distro-info==0.18
entrypoints==0.3
httplib2==0.9.2
idna==2.8
idna-ssl==1.1.0
importlib-metadata==0.19
ipykernel==5.1.1
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.14.1
Jinja2==2.10.1
jsonschema==3.0.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==7.2.0
multidict==4.5.2
mypy==0.670
mypy-extensions==0.4.1
nbconvert==5.5.0
nbformat==4.4.0
netifaces==0.10.4
notebook==6.0.0
numpy==1.17.0
oauth==1.0.1
olefile==0.45.1
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.1.0
pluggy==0.12.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.0.0
ptyprocess==0.6.0
py==1.8.0
pycairo==1.16.2
pycares==3.0.0
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.4.2
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.2
pyRFC3339==1.0
pyrsistent==0.15.4
pytest==4.2.0
python-apt==1.6.3+ubuntu1
python-dateutil==2.8.0
python-debian==0.1.32
python-dotenv==0.10.3
pytz==2019.2
pyxdg==0.25
PyYAML==3.12
pyzmq==18.0.2
qtconsole==4.5.2
reportlab==3.4.0
requests==2.22.0
requests-toolbelt==0.9.1
requests-unixsocket==0.1.5
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.13.2
six==1.12.0
system-service==0.3
systemd-python==234
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
typed-ast==1.3.5
typing-extensions==3.7.4
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.25.3
usb-creator==0.3.3
virtualenv==16.4.0
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
xkit==0.0.0
yarl==1.1.0
zipp==0.5.2
zope.interface==4.3.2